model_api/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
diskcache==5.6.3
fastapi==0.115.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
llama_cpp_python==0.3.0
MarkupSafe==2.1.5
numpy==2.1.1
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
